Create a base cell for audio cell containing all the setup code to be called in the subclasses
